some refactoring added tests for MultiItemPricingRule

+++ src/main/java/com/itv/airtime_sales/checkout_kata/Unit.java
@@ -3,9 +3,9 @@ package com.itv.airtime_sales.checkout_kata;
 public class Unit {
 
     private final long price;
-    private final long remainder;
+    private final int remainder;
 
-    public Unit(final long price, final long remainder) {
+    public Unit(final long price, final int remainder) {
         this.price = price;
         this.remainder = remainder;
     }
@@ -14,7 +14,7 @@ public class Unit {
         return price;
     }
 
-    public long getRemainder() {
+    public int getRemainder() {
         return remainder;
     }
